Week9/requirements.txt
python-dotenv
pandas
faiss-cpu
sentence-transformers
autogen-agentchat
autogen-core
autogen-ext
asyncio
autogen-ext[llama-cpp]
autogen-ext[ollama]
autogen-ext[openai]
